permitir borrar solo los datos que no se esten usando
se a colocado en todas las tablas excepto a la tabla empresas, una revision antes de borrar para que si el dato a borrar se esta usando no se pueda borrar y si el dato no se esta usando se puede borrar.

+++ frontend/finalproject/src/components/Droides/Droides.jsx
@@ -22,16 +22,28 @@ function Droides() {
         API.getDroides().then(setDroides)
     }, []);
 
+    // mensajes de advertencia
+    const [mensaje, setmensaje] = useState('');
+
     // borrar los datos
     const borrarDroide = async (event, id_Droide) => {
 
         event.preventDefault();
 
-        // eslint-disable-next-line no-unused-vars
-        const request = await API.deleteDroides(id_Droide);
+        const request = await API.postDroidesAndEmpresasByID(id_Droide);
 
-        setInterval("location.reload()",1000);
+        console.log('el dato se puede borrar -- ',mensaje);
 
+        if(request.status){
+            setmensaje(request.mensaje);
+            const borrer = await API.deleteDroides(id_Droide);
+            setInterval("location.reload()",1000);
+            alert(mensaje);
+
+        }else{
+            setmensaje(request.mensaje);     
+            alert(mensaje);
+        }
     }
     // verificar usuarios
 
@@ -83,8 +95,6 @@ function Droides() {
                 </div>
 
                 <div className='containeButtonCargar'>
-
-  
                     <Link to={`/AgregarDroides/${usuario}`}>
                     <Button variant="dark">Cargar Droide</Button>
                     </Link>
